87. User enabled property migration and fixtures(Section 13: API Platform - Custom Resource, E-mail Sending (User))
In this section we will create a system that requires users to confirm their account after they register.
So we will be sending an email with a link that contains a link that can confirm a user with a specific.
Create the specific fields and enable all users in fixtures.