91. UserConfirmation custom API Resource(Section 13: API Platform - Custom Resource, E-mail Sending (User))
Now that we have created the confirmation talken for each user that registers will have to somehow be

able to accept this confirmation tokens that users sent to us and last time we have create that custom

operation for our user entity.

But this time we want to create a custom operation we will create our custom API resource and this resource

will be different than all the others that we have.

So you will see the inside API platform to create a resource you just need a plain old pupae class.

It does not have to be an entity.

It makes things a lot easier when you have to persist data and you want to use that built in doctorin

support.

But every class can be API rissoles and you can have some individual ways of storing this data or just

have some action that responds to certain operation.

In our case this will be only one operation.

So this user confirmation class will only have a post-operation that will accept the confirmation token

that the user has been generated.