100. Image entity with migration and @Uploadable annotation(Section 14: API Platform - Uploading Files (Images))
Now if you would like to upload more than one type of file first three name our products the images

because we have only one file type to upload but in case you will need for example specific another

type of image just like profile pictures you can add another mapping.

And define older properties like Eury prefix the upload destination and the namer.

You can also upload different kinds of documents for example you might want to upload PTF.

So for each of those files they will require another NTD you can define those mappings here.

So since we only have general images let's create a new image A.D. and this time we'll do that manually

instead of using the generator.

So create a new image class inside our NTD and mind that when you are creating this image class manually

You will also have to create a repository class if you need one.

You will have to do that manually now by simply adding the entity annotation on top of the image we

are marking this class as a doctrine entity so now it will be managed the by doctrine automatically

and we will be able to generate migration for this class.

So by using their anti-Tea notation the page B-story automatically imported that doctoring or mapping

name namespace using their or Elias.

So if you didn't have a huge storm you will do that manually.

Now we need an ID.

So each of the entities that we have needs some kind of unique ID and this would be this has to be marked

with the ID annotation and the generated value to automatically increment the next ID fields.

Now we will need a r l that this will contain the actual you up to this image and the first field that

we need is that file.

So file is not actually a field that we will store inside the doctorin so you can call it a virtual

field.

And now we'll have to upload a ball field notation from the Veach uploader Bondel and by simply adding

this uploader ball field annotation on our property the Veach uploader Bondel will be able to handle

file uploads for us.

So we need to specify a property which is then mapping.

So that's why we defined our mappings.

And we currently have only one of course but if you will have more than one mapping then now is the

time to specify that for our image anti-Tea it should use their mapping called images so that has to

be one of the mappings that you have defined inside the configuration file and the upload of uploader

ball field requires another setting which uses that file name property which would be the user out to

that image which clients could use to actually see that image.

So it is the user field.

Now we also have to mark this entity with the Veach up loadable notation.

So the average uploader bundle can pick this up and handle the uploads for us in the background.

Now you will also need to generate getters and setters.

This is what we usually do as our fields are private.

So let's generate getters and setters for our you are real file field and ID.

And of course as usual we do not use Surtur for ID because it is bang ID is being set by doctorin only

when this entity is being persistent.

Now finally we have to mark this image entity with an API resource so it will be out magically available

inside our API and we'll be able to upload files.

We will get to that later.

Now we also need to make immigration so run ph be being Konsole make migration to create new migration

for our new entity and then run it using doctorin migrations migrate.

So it should create our new images anti-Tea and will continue working on that in the next lecture.