101. API Resource for Image entity{set _api_receive}(Section 14: API Platform - Uploading Files (Images))
There is a setting called _api_receive and set it to false.

This will just forbit API platform of creating an instance of this class automatically for

us from the request body.

And this makes sense because we will be uploading an image and handling this work to this custom operation.