102. Implementing custom Action for upload(Section 14: API Platform - Uploading Files (Images))
We have to call the form factory create method and the first argument of this is the form type.

So you will have to create a form type class that will actually define what is inside this form.

We will do that later.

So if everything is fine we will persist our new entity inside the database and we will just return

this new entity back to the user and then API platform will handle serializing it back for the user.

But first will have to set that file to null.

And this is because this will contain the binary file that we have uploaded and when we will return

the instance of image with this file assigned this will produce a massive response that will also contain

the binary of the file inside this file image properties.

So this really wouldn't look very nice.

And this is actually useless.

So that's why in-case their form was involved we are injecting the validator component and thats why

we are calling the validation again because if we would send the file to null at this point and just

return the image Well if we all tried to leave the foundation to the API platform it will always show

that there was an error because the only validation constraint that we have is actually on the file

field.

And if we will be setting it to know there variation will always fail.

So thats why we are doing a trick like this.

We are following the validation exception immediately with all the errors that at this point we are

certain that the file field was no for the other reason that we have set it to nil.

So the user just didn't provide it.

So that's more work around that we have to do.