103. Creating Form for file upload(Section 14: API Platform - Uploading Files (Images))
The get block prefix is required and will only return an empty string because it only makes sense
when you are actually rendering a form in a traditional Symphony controller.
And we are not doing that.
We do not render this form visually.
And the required field is only used to our HD and HTML5 validation.
Our foreign class is not a typical form. We only used to be able to send the request as form.
Now let's see what was our route for uploading files.

And this was post API images

so since we are handling this as a form inside body you won't choose the wrong option.

You will choose the file.

So that's the only field our form has it's the file field.

And inside those field we will select a file that we would like to upload.

So choose the file option from the available options and well you can use an image for uploading.

Maybe let's use a smaller image.

So because by default PHP will have an upload limit of two megabytes.