108. Date filter(Section 15: API Platform - Collections Filtering &amp; Sorting)
In this lecture we will add a brand new filter to our blog posts.

And this time we will see how we can fetch parts of our collection by filtering using the publish date.

So for comparing dates we have to define another filter the search filter won't work because it's just

literally matching the values of the parameters.

And this time we'll need a lot more options so let's add a new API filter annotation and as its first

parameter specified a date filter class.

Remember that you will have to import that class from the namespace that you see on the top and we use

that data filter to use some to do some complex worries on the publish date.

So we'll be able to find if this bobbish date was before a certain date or if it was after a certain

date we'll be also able to search for date ranges.

When our blog post was published.

So let's start with the simplest query and we will do that by adding their property name to our query

and doing an array so surrounded with square brackets and inside the square brackets you will specify

that parameter that you will find it by.

In this case we will be finding blog posts that were published after the specific date.

So this will find everything that has that date which is bigger than 12 of February 2000 18.

We can see that we have forty eight of those items.

You might have more or less depending on how your fixtures are generated because the dates are random.

So after the seventh month of 2018 you can see that there are only 4 items.

Let's try to add another query so you can add more than one condition and use using an operator.

This time we would like blog posts published after 12 but before 20 if

so there shouldn't be a lot of those block posts.

We can maybe modify it to the first tee and let's see how many blog posts we still have free blog posts

so we are actually running out of the blog posts.

You can also try to be more specific since our Published Date contains not only the date but it also

contains the time.

So let's try to create a more specific filter and see if after the filter works so we have specified

that we want to find something after 16 but the time is actually passed so it is after.

Now you should know that when using the after filter it will also include the date that you have specified

as an argument but if you specify it strictly after it will only work for dates that are actually after

your specific date so it will not include the date that you put inside the filter and the same works

for strictly before.

So after will include the date inside the parameter and strictly want it will just work for everything

that's strictly after date as you can see we specified the specific time of the of that block post and

it wasn't included after we modified the strictly after just after it was catched.

Using this query parameter.