109. Range filter 110. Sorting filter(Section 15: API Platform - Collections Filtering &amp; Sorting)
In this lecture we will learn about another filter that is out there.

It is called range filter.

So it allows us to filter certain values using the range of properties so you can specify that it should

be for example greater done something or less than something we'll use that for ID because that's the

only property that makes sense in our case.

So you will be able to compare the ID to certain values.

Let's start with that greater than which is done using DDT.

So G.T. stands for greater than and well that's what happens if you do not import that new range filter

class because we have specified it inside the annotations so we have to import it.

Now let's see after this release now it actually works and you have as you can see that we have found

sixty five items that actually in my case had ID greater than one thousand four hundred fifty.

You can have different IDs depending on how many times you have loaded the fixtures.

Now if you were fully used to conditions greater than and less done.

So can combine that.

We want everything that is greater than 1 4 5 0 0 and everything that is less than 1 4 6 0 so are much

more or less items.

You can also use DDT which is greater than or equal or LTE which is less than or equal.

110. Sorting filter

In the first lecture of that section we have seen how can we set the default ordering of our collections.

Now let's see how we can let users decide by what properties and using what order they would like to

sort the results that they see when the fetching their collections.

So you'll have to add another API filter and use the order filter class as its first argument.

You will have to also import that class so it's very important to always really remember about it.

And inside the properties setting of those API philtre you specify by what properties you will allow

users to sort the results.

We have made a mistake so I forgot one column.

That's why we get error after refreshing the results.

It now looks correct.

So let's try to sort this collection by one property and this could be used.

This could be to order by ID.

So we will specify the sorting properties by specifying an order inside query and inside the square

brackets you will specify by what property you want it sorted and the value of that parameter would

be the sorting order.

So we are now sorting the results of by ID using ascending order if you'd like to add another parameter

to the sorting Kohary you will have to add another order property and in the square brackets specified

the property of that resource.

If we would like to sort of by the publish date we'll have to put publish date inside square brackets

and the value of it should be by what order we would like to sort.

So if you would like to see the oldest posts first you should specify that the order should be the publish

date and the sorting order should be ascending so A.S.C. barometer says that it will be sorted by the

oldest posts first.

Now you can also specify the name of this ordering parameter by default it is order if you would like

it different then you can add additional property inside the notation.

And we have modified that to underscore order and you can see we are confirming that now by specifying

underscore order we have our sorting parameter.

And if you modify it to order and we have already modified how we want our property to look like you

can see that it is using the default sorting so that arguments order parameter name is the way to specify

how you would like to order your collections or how you would like to let users order your collections.

Now by changing this parameter inside the entity definition we only modified the name of the ordering

parameter for that particular entity.

If you want to change it globally go to the API platform Yarm a configuration file and change the order

parameter name under collection.

We want to save that but this is the way you will do that if you would like to have a different name.

For example in your own language.