112. Property filter(Section 15: API Platform - Collections Filtering &amp; Sorting)
Now let's see the last filter that I'd like to show you.

It is called the property filter.

It's different than the other filters because now you won't be filtering the data that you will receive.

So you are not to limit the data that you would like to get.

You will only limit what properties of this particular entity you would like to be returned.

So imagine you do need everything about the block posts.

You only need title.

So by adding a new API field or with that property filter class you'll be able to configure it using

arguments and specify the parameter name the parameter name is the parameter you will use.

Inside the query string to say or to signal to API platform what properties you would like to return.

There is also another option called override default properties and if you were to specify it as true

you will be able to override whatever configurations you have used using the serialization groups so

if some property is not inside the serialization groups you can anyway be able to fetch that property

if you will to set override default properties to true.

So users will be able to specify that they want to see that property.

So watch out with this setting.

And the final setting is the white least you'll specify older properties that you would like to allow

the clients to see.

So for example if you'll have a list of users you might not include the email inside those white list

because we do not want to show user emails to everyone.

So even if someone would specify that they want e-mail in the query parameter they wouldn't be able

to see it up.

So we had an error and probably that's our usual case we forgot one characters so you need to use double

column ins instead of one.

Now what are we doing here.

We are specifying the properties as we wanted it in to have that name.

And we have added ID to the array.

Now we can also set that accepted format that we want this to be returned so if you want only the data

without any major data you can accept that you only expect application and Jason.

So the default return type is application L.D. Jason which also returns all the Hydra METAR parameters.

So if you want just the straight data and those three properties you can also set accept hether to be

application slash.

Jason.