113.1. Configuring collection pagination(Section 16: API Platform - Collections Data Pagination)
In this lecture we will try to see how we can modify the way the pagination works inside a platform

so we can see that by default we have some meta data Hydra Tottle items which tells us what is the total

amount of items of that particular resource.

And it also provides links to particular pages like the next page.

The last page and the first page for example.

So when you will click on those links you can see that we can view the resources on another page.

And this is all inside API platform by default but it also has some Cuppy capabilities to configure

the way it works.

So you can change the page parameter name by default.

It is just a page but we will be now configuring it globally.

So instead of page this will be underscore page and it is ultimately used by API platform in all resources.

So we have configured that globally.

You can use that for example if you would like to have that parameter insert in your own language like

German French.

You can use any other word.

Now you can also disable pagination and we just did that globally.

So all the items are in one single big list.

So this might be useful if you have some parameters or you have some resources that you always want

to show in a big list.

This might make sense if you have kind of resource that is not the least of those resources is not so

big.

So you have for example 20 resources of some kind and you do not need Puggy nation on that because you

would like to read them all at one request.

You can also control how pagination works per individual resource.

For example we can disable pagination only for comments.

So by adding to the attributes of this API resource that pagination enabled equals force will be able

to disable pagination only for their comments.