113.2. Configuring collection pagination(Section 16: API Platform - Collections Data Pagination)
If you will if you will enter that sub resource comments for our particular blog posts you can see that

all the comments are inside.

But basically since we by default.

Pagi neighed by 10 you actually can't see that so you'll have to find a block post that has more than

10 comments to see this at work.

So you can also try it out yourself.

And for example temporarily disable packy nation for blog posts only for the blog posts so you will

see that they will be all on one page after that.

Now you can also specify additional parameters that the user is Quadrilogy or collections will be able

to decide if he wants those collections Bagian it or not.

So by adding those global properties you are able to specify whether you want pagination or not.

So by adding up our ometer pagination equals false we can disable the buggy nation forward to block

posts so all the items are displayed inside this one page.

Now if you will modify to true you can see that now the resources are being stagnated.

So you see the meta data for all the individual pages inside the blog posts so client enabled to let

their consumers decide and enable parameter name to specify in the query parameter whether the clients

want the pagination or not.