113.3. Configuring collection pagination(Section 16: API Platform - Collections Data Pagination)
This can be also configured per resource.

So lets modify the attributes inside comment and add an attribute pagination client enabled to true.

And this way we'll be able to let the consumers of our API decide whether they want buggy nation on

comments or whether they don't want the pagination on comments.

So if we have disabled this by default now the clients might enable it using that pagination parameter.

Since this particular example has only one page it actually didn't work.

So the page it anyway showed only one page.

So there was nothing to paginate.

But if you will have resources that have more items than just 6 then you will be able to see that it

actually works.

And that's the way you can let users decide whether they want their PAGENAUD resolved or not.

And by users I mean consumers of the API because you might be creating a front end application that

will be only a user of those API but you might also have this API for different consumers.

So let's see how we can configure item sper page since it is 10 by default.

We now specified it to be 4.

So we should see the page but we have disabled the pagination using parameter and there were no pages.

Now when we specified that we want the pagination you can now see that we have two pages inside comments

because we have said that there should be four items per page.