114. Configuring collection pagination part 2(Section 16: API Platform - Collections Data Pagination)
Let's continue exploring what are our other options of configuring pagi nation for our resources collections.

So in this example with the comment resource we have the pagination client enabled set to True which

means that the consumers of our API can decide if they want this results to be pagen ated or not.

So let's add another property and this time it is that Diggnation client items per page.

So now we will let clients decide how many items they would like to see per page.

So the default is then we can always change that inside API platform Yunel collection machinations items

per page and there is also generic property client items per page which is disabled by default and you

can all in this place you can enable it for every resource so the consumers of the API can decide how

many items they want and also those items per page parameter name which would be to represent what is

the query parameter that you would like to use to specify the number of items per page that you would

like to see.

So let's move back to the postman and see how we can individualize what we got back from the API.

So lets parameter items per page 2 and we will signal to the API that we want comments to be pugging

ated the way that we see two items per page.

So we can see that there are six items in this collection in this collection.

The Hydra total items says that it has six items and we have.

And we had free pages.

Now if you will specify that you want five items per page you can see that we had two pages

and now inside the block posts you can see that we have hundred block posts so there are 10 pages and

if you would like to have 50 items per page you can see that we still get the 10 items because we have

to first enable that to to the customers that they can specify how many pages they would like to see

on an individual page.

So it is disabled by default and we have also disabled that in the country.

So if would like to let those blog post consumers decide how many they want on one page of blog posts

you will have to either to directly inside the resource or enable it globally.

Now we can see that the blog posts.

It was enabled for everyone.

So we saw that we have hundreds of blog posts on one page.

You can also specify the maximum items users can specify so in case users would like to see hundred

items on one page and we do not want to allow them to do that we can specify maximum items per page

and this value will override anything that consumers would specify.

So if they would like a hundred items and you let them see only 20 items thats the way to do it you

can also specify this property per resource so you can say that generally you allow 20 items per page

but for example for the blog posts you make an exception and you allow users to see 40 items per page

so if you see words the first page and last page this would make sense.

The last page is page four because if you will divide hundred by 40 this would be four.

So we have four pages so this is the way you can configure pagination even further.