115.1. Partial pagination (performance gain)(Section 16: API Platform - Collections Data Pagination)
Now the last parameter of pagination that I'd like to show you is that partial pagination.

So first lets enable that globally for every resource.

And what this would change is that you will only have a link to the next page.

So the API won't return how many items are there in total and what's the last page.

It would only be the next page.

So this is a performance optimization because it doesn't have to do a count query on all the items.

So if you for example have a use case where you browse items one by one let's compare it to our Facebook

wall.

When you scroll you just load the next items when you scroll.

You do not need to know how many posts are left for you to see because that might be endless So in cases

where you only have partial nation this is a great optimization.

You can also define that per resource.