115.2. Partial pagination (performance gain)(Section 16: API Platform - Collections Data Pagination)
So this probably wouldn't make sense to define that partial pagination globally for everyone.

You would be only doing that for research says that might be view one after one.

You can also lead Kline's the site that they want to browse resources partially and by adding their

client partial to true and partial parameter name you let users add this partial parameter and set it

to true.

So they won't see the total number of items and the total pages they will just see the link to the next

page.