118. Fixing validation groups(Section 17: API Platform - Error Handling, Logging)
In the last lecture we've notice that when sending empty JSM we actually see an error from the database

so it's like our validators do not work at all.

And one of the reasons that we sit up after making changes is that we do not have tests yet and the

exact reason is that we have defined groups for our validation constraints.

So for for example the username and password should only validate when the validation group is post.

So it's not enough to said that then or maybe zation context.

You have to also said what's that validation group for individual requests request if you use the validation

groups so it's not the same thing.

It's not that you said the group property for a particular request.

And it works for the serializer and the validator.

They are separate.

So for our posts request we have to set the validation groups property and set it to post because that's

how we have called our validation groups inside validation constraints.

You will have to do the same for the pute reset password because we have different validation groups

for this.

We have the validation groups called Putu reset password.

So you can call it a coincidence that our validation and Serializer groups are called the same.

I tried to do this for simplicity but those groups are separate so by if you specify groups for validation

you cannot just set the serialization group of a request.

It has to be also specified inside the validation groups property.