119. Handling business logic exceptions(Section 17: API Platform - Error Handling, Logging)
So last time we have added new exception to handle or case we're on empty bodies being sent through

post or pood requests.

So in this scenario where we send the data to the backend.

Let's see another example where a custom exception might be useful.

And eat this our confirmation token confirmation logic.

So when we will confirm our talkin when we try to confirm Otakon and the user wasn't found by disconfirmation

talkin we were following not found HTP exception.

This was enough for then and this produced an error message have 404 status code which is a standard

status code for when something could not be found.

So whenever your API gets asked to access resources that do not exist you would return for a hundred

for error.

Now that message that something couldn't be found may be too general and we can handle it better.

So let's create a new class and call it an invalid confirmation token exception.

And this would extend the base exception class.

Now we can now follow this exception inside our business logic because the user confutation servies

is a generic service that you will be able to use even outside the API.

So this can be called for example Frue command line commands.

So we are extending our base exception class so it doesn't have any status code inside it or anything

related to the request itself and we'll configure it to return 400.

For starters code.

Now let's test that out using our code token confirmation and point.

So first we have this validation errors and this is correct because this entity is being validated by

API platform.

The same feeling when the confirmation token is not exactly for the characters as we specified it to

be.

So let's now try to make it exactly 40 characters and you see that while this stock did not exist in

the database and now we have our 404 error we have our custom error message.

You can also change it to 400 if you prefer.

Anyway you can also leave it as a default message that IDM or resource just could not be found.

This would make perfect sense because in case of when we check anything related with security it's not

a good idea to give out too much information so you shouldn't be too specific of what is incorrect what

should be fixed because someone who might use this information to exploit something inside your system.

But this is just an example of how we can have exceptions inside our business logic that is basically

outside the request response lifecycle.

And in this case it is our service class and how you can convert that to meaningful messages returned

straight from the API.