120. Installing Monolog library for application logging(Section 17: API Platform - Error Handling, Logging)
In this lecture we will see how we can log error messages or debugging messages inside our API.

So let's start by installing the most popular logging library which is then monologue and we can install

prepared Bondel for the monologue using composer require sinfulness slash monologue bundle.

So this will stop the monologue bundled for us and initially configure it after installing it.

We will take a quick look of different configuration options that it provides and the default config

that is created for us.

So as you can see we have free monologue yum configuration files and they are both different for different

environments.

As this makes sense because you have different levels of aerostat you would like to log on different

environments for example on the development environment.

You might lock everything because you are developing.

You would like to see even the debug messages so not only those that are pretty critical but on the

other hand on the production environment you probably would only like to log their critical critical

messages and maybe also everything that led up to them.

Now let's take a quick look on the default configuration file for that dev environment and you can see

that we have two handlers defined so you will define handler each time you would like to specify what

should be stored where it should be stored and when it should be stored.

So let's start with the handler called Main.

By the way you can call the handlers however you like it's up to you.

So the main handler has a type stream stream means it the locks will be stored inside files next that

BAAF option is to specify where those files should be stored.

It's filename pattern for the files that will contain the locks.

So in this case is the code in the logs dir by default that would be var slash log and the file name

will also contain that the environment that produced that error.

Next we have lever.

So it can define different levels of where when that logger should actually log into those files.

So you can have debug error are critical.

So it's up to you when you would like to start storing them as such.

So everything that is less important will not be stored.

For example with this setting that the bug everything that will contain that info level would not be

stored.

And the fourth option is channels and using channels you can specify which types of messages you would

like to store in using this handler inside file.

And there are some default channel was created and configured by Sinfonie That's for example doctorin

event security on request.

So they are used to categorize the log messages inside Symphonie.

In this particular example we have an asterisk before an event.

So instead of defining what kind of events we would like to store we find that we want to store everything

except the messages from that event channel.

There is also a second handler defined the console which will output them messages to the console and

you can check that out yourself if you will run the built in BHP server and the way everything works

here is the sinfulness or monologue.

We'll just use every handler in the order they are defined.

So first the main handler then the console handler and it will try to see if those settings match to

the messages that should be logged and then we'll use that handler to produce the output.

There are exceptions though sometimes you might want to additionally process some messages or on the

user handler that will be a kind of processing the message messages doing something with them before

it actually gets passed to be published or out put it somewhere.

But we'll get to that later.

Now let's make a request to our confirmation talk and using that data that you can see and then we'll

check the locks and see what got inside the locks so you can find the locks from stream inside that

def log file and you can see that every message is being logged using that channel which in this case

is security.

And after that dot there is the level of that logs so that info debark critical depending on what happens

sometimes you might just log that something happened there where it was made and sometimes obviously

you logged the error messages that something unexpected has happened.

Let's see how we can lock something explicitly lets other logger in their face as our dependency to

our user confirmation service and by default.

Sinful man has a service called lager.

And if you install the monologue Bondel it'll be on the AS to the monologue to the monologue logger.

Now the logger has couple of metals and they match the level of severity of what happened.

So we might want to lock that during the confirmation process something is happening.

So we'll be using the debug level and logging that we are fetching the user by confutation token then

if the user was not found We'll lock that user by consumation token has not been found not used up.

It's not actually an error or we are just logging that someone tried to confirm a token and maybe that

token was expired.

So we can put that all inside our lock files and may be the final message would be that confirmation

was successful.

So maybe let's at least try to provide our rank talkin but let it be something.

Now you see that we have these messages and they are by default using the app Johno.

So everything you log explicitly from your app has an up channel and that level is correct.

So it is our debug level and you can see our exact messages that we have created.